Extract current conditions card
