+++ src/components/LocationSelector.tsx
@@ -5,11 +5,9 @@ import {
 	CheckCircle,
 	Cloud,
 	Edit,
-	Info,
 } from "lucide-react";
 import { haptic } from "ios-haptics";
 import { useAppStore } from "../store";
-import { useSweatIndex } from "../hooks/useSweatIndex";
 import {
 	getCurrentPosition,
 	reverseGeocode,
@@ -17,9 +15,8 @@ import {
 } from "../services/geolocation";
 import { fetchWeatherData } from "../services/weather";
 import { LocationSearch } from "./LocationSearch";
+import { CurrentConditionsCard } from "./CurrentConditionsCard";
 import type { GeocodingResult } from "../services/geocoding";
-import { formatTemperature } from "../utils/temperature";
-import { getWeatherIconDetails } from "../utils/weather-icon";
 import { Card, CardContent } from "./ui/card";
 import { Button } from "./ui/button";
 import { Alert, AlertDescription } from "./ui/alert";
@@ -32,11 +29,6 @@ export function LocationSelector() {
 		setWeather,
 		setGeolocationError,
 	} = useAppStore();
-	const sweatIndex = useSweatIndex();
-	const currentWeather = geolocation.weather?.current.weather[0];
-	const currentDewPoint = geolocation.weather?.current.dewPoint;
-	const weatherIcon = getWeatherIconDetails(currentWeather?.id);
-	const WeatherIcon = weatherIcon.Icon;
 
 	const handleSearchSelect = async (result: GeocodingResult) => {
 		try {
@@ -221,86 +213,7 @@ export function LocationSelector() {
 						</div>
 
 						{geolocation.weather && (
-							<Card>
-								<CardContent className="p-4">
-									<div className="flex flex-col gap-4 sm:flex-row sm:items-start sm:justify-between">
-										<div className="flex items-center space-x-4">
-											<WeatherIcon
-												className={`w-8 h-8 shrink-0 ${weatherIcon.className}`}
-											/>
-											<div>
-												<p className="text-sm text-muted-foreground">
-													Current Weather
-												</p>
-												<p className="text-sm font-medium capitalize">
-													{currentWeather?.description || "Clear"}
-												</p>
-											</div>
-										</div>
-										<div className="grid grid-cols-2 gap-3 sm:min-w-72">
-											<div>
-												<p className="text-xs font-medium text-muted-foreground">
-													Temp
-												</p>
-												<p className="text-xl font-bold tabular-nums">
-													{formatTemperature(
-														geolocation.weather.current.temp,
-														geolocation.weather.temperatureUnit,
-													)}
-												</p>
-											</div>
-											{currentDewPoint !== undefined && (
-												<div>
-													<p className="text-xs font-medium text-muted-foreground">
-														Dew point
-													</p>
-													<p className="text-xl font-bold tabular-nums">
-														{formatTemperature(
-															currentDewPoint,
-															geolocation.weather.temperatureUnit,
-														)}
-													</p>
-												</div>
-											)}
-											<div>
-												<p className="text-xs font-medium text-muted-foreground">
-													UV Index
-												</p>
-												<p className="text-xl font-bold tabular-nums">
-													{geolocation.weather.current.uvi}
-												</p>
-											</div>
-											{sweatIndex && (
-												<div>
-													<div className="flex items-center gap-1">
-														<p className="text-xs font-medium text-muted-foreground">
-															Sweat Index
-														</p>
-														<Info
-															className="h-3.5 w-3.5 text-muted-foreground"
-															aria-label="Sweat Index is an outdoor comfort score based on temperature and dew point in Fahrenheit. Higher values mean hotter, muggier air."
-														>
-															<title>
-																Sweat Index is an outdoor comfort score based on
-																temperature and dew point in Fahrenheit. Higher
-																values mean hotter, muggier air.
-															</title>
-														</Info>
-													</div>
-													<p className="text-xl font-bold tabular-nums">
-														{sweatIndex.value}
-													</p>
-													<p
-														className={`text-xs font-medium ${sweatIndex.textClassName}`}
-													>
-														{sweatIndex.description}
-													</p>
-												</div>
-											)}
-										</div>
-									</div>
-								</CardContent>
-							</Card>
+							<CurrentConditionsCard weather={geolocation.weather} />
 						)}
 					</div>
 				);
